Implement add new contact modal
